+++ VMWithEditor/UnitTests/Parser_Unit_Test/Parser_Unit_Test/parser_unit_test.h
@@ -1,6 +1,6 @@
 #ifndef PARSER_UNIT_TEST_H
 #define PARSER_UNIT_TEST_H
 
-bool unit_test_parser(char* file_name, unsigned test_step);
+bool unit_test_parser(char* file_name, size_t test_step);
 
 #endif
